+++ filament-core/src/androidNativeMain/kotlin/dev/karmakrafts/filament/Thread.kt
@@ -108,7 +108,7 @@ internal actual fun getThreadId(): ULong {
 }
 
 @OptIn(UnsafeNumber::class, ExperimentalForeignApi::class)
-internal actual fun suspendThread(millis: Long): Long = memScoped {
+internal actual fun sleepThread(millis: Long): Long = memScoped {
     val spec = alloc<timespec> {
         tv_sec = (millis / 1000).convert()
         tv_nsec = (millis % 1000000).convert()
@@ -129,6 +129,6 @@ internal actual fun isThreadDetached(handle: ThreadHandle): Boolean {
     return handle.value in detachedThreads
 }
 
-internal actual fun setThreadAffinity(vararg logicalCores: Int) {
+internal actual fun setThreadAffinity(vararg logicalCores: Int) = Unit
 
-}
+internal actual fun getThreadAffinity(): IntArray = intArrayOf()