+++ filament-core/src/appleMain/kotlin/dev/karmakrafts/filament/Thread.kt
@@ -115,7 +115,7 @@ internal actual fun getThreadId(): ULong = memScoped {
 }
 
 @OptIn(UnsafeNumber::class, ExperimentalForeignApi::class)
-internal actual fun suspendThread(millis: Long): Long = memScoped {
+internal actual fun sleepThread(millis: Long): Long = memScoped {
     val spec = alloc<timespec> {
         tv_sec = (millis / 1000).convert()
         tv_nsec = (millis % 1000000).convert()
@@ -137,3 +137,5 @@ internal actual fun isThreadDetached(handle: ThreadHandle): Boolean {
 }
 
 internal actual fun setThreadAffinity(vararg logicalCores: Int) = Unit
+
+internal actual fun getThreadAffinity(): IntArray = intArrayOf()